Rename namespace
See https://github.com/NationalSecurityAgency/ghidra/issues/7
