+++ Ghidra/Framework/Docking/src/main/java/docking/framework/DockingApplicationLayout.java
@@ -23,8 +23,8 @@ import java.util.Objects;
 import generic.jar.ResourceFile;
 import ghidra.framework.ApplicationProperties;
 import ghidra.util.SystemUtilities;
-import utility.applicaiton.ApplicationLayout;
-import utility.applicaiton.ApplicationUtilities;
+import utility.application.ApplicationLayout;
+import utility.application.ApplicationUtilities;
 import utility.module.ModuleUtilities;
 
 /**
