+++ Ghidra/Framework/Utility/src/main/java/utility/application/ApplicationLayout.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package utility.applicaiton;
+package utility.application;
 
 import java.io.File;
 import java.io.IOException;
