+++ Ghidra/Framework/Utility/src/main/java/utility/application/ApplicationSettings.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package utility.applicaiton;
+package utility.application;
 
 import ghidra.framework.PluggableServiceRegistry;
