+++ Ghidra/Framework/Utility/src/main/java/utility/application/ApplicationUtilities.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package utility.applicaiton;
+package utility.application;
 
 import java.io.*;
 import java.util.ArrayList;
